SIM-1308 ADD methods to fixture for adding generic arguments and formal parameters

+++ tests/sample/src/test/java/eu/jsparrow/sample/postRule/allRules/TestPrimitiveObjectUseEqualsRule.java
@@ -155,7 +155,7 @@ public class TestPrimitiveObjectUseEqualsRule {
 		}
 		if (d.equals((Integer) c)) {
 		}
-		if (c == (Integer) d) {
+		if (c == d) {
 		}
 	}
 }
